+++ Client/SineGeneratorForInt64.cpp
@@ -13,7 +13,7 @@ SineGeneratorForInt64::SineGeneratorForInt64(qint32 bytes) {
     qint64 sineValueForQByteArray = 0;
 
     for (int i = 0; i < 1000; ++i) {
-        sineValueForQByteArray = qSin(i * 2 * M_PI / 1000) * INT16_MAX;
+        sineValueForQByteArray = qSin(i * 2 * M_PI / 1000) * INT64_MAX;
         m_block.append(reinterpret_cast<char*>(&sineValueForQByteArray), size);
     }
 }
